fixed some bugs, added console log for debugging

+++ socketEvents/notificationDismiss.js
@@ -5,7 +5,8 @@ module.exports = async (data, client, io) => {
     if (!channelID) return; 
     
     const { ok, result, error } = await redis.getConnectedBySocketID(client.id);
-
+    if (!ok || !result) return;
+    
     const uniqueID = result.u_id
     await Notifications.deleteOne({recipient: uniqueID, channelID});
     io.to(uniqueID).emit('notification:dismiss', {channelID});
